Figured out how to distribute the Android PK jars inside the plugin.
The two jar files come directly from the AndroidProximityLibrary distribution.
However, I had to rename `classes.jar` to `AndroidProximityLibrary.jar` to
be safe and explicit.

This obviates having to add the AndroidProximityLibrary as an Android
library project by hand directly to the application project.  It does this by
modifying the app's `AndroidManifest.xml` and adding the `<uses-permission>` and
`<service>` elements needed.  This is accomplished by the plugin's `plugin.xml`
so no additional sleight-of-hand is taking place.