Convert watchCallbacks from NSMutableSet to NSMutableDictionary
Previously watchCallbacks was an NSMutableSet that only stored the callbackId as a string for each watch created by the plugin.  Because of this, there was no way to tie the watchId (needed to cancel callbacks with the clearWatch method) to the corresponding callbackId.  By storing the callbackId in a dictionary keyed by the watchId, clearWatch can now cancel callbacks using just the watchId.